Django_Svelte/requirements.txt
coderedcms==0.20.*

# django_sendmail_backend enables sending email from your web host server.
# Remove this if using a different email backend.
django_sendmail_backend

# To use with MariaDB or MySQL, uncomment the line below:
#mysqlclient
# To use with PostgreSQL, uncomment the line below:
psycopg2
